touching and jumping
Previously if you hit a wall (e.g. on your right) but held down right
and then jumped it would jump purely vertically because your right
movement was stopped.

Now the action of touching the wall (continously) doesn't stop the right
action, but resets the x0 and the xstart timestamp, so if you jump then
it will jump upwards until you're not touching the wall any more and
then start moving right.